add size_t casting for clarity

+++ bitarray/_bitarray.c
@@ -511,7 +511,7 @@ count(bitarrayobject *self, Py_ssize_t a, Py_ssize_t b)
         if (m) {                         /* starting at p count in m bytes */
             uint64_t tmp = 0;
             /* copy bytes we want to count into tmp word */
-            memcpy((char *) &tmp, self->ob_item + p, m);
+            memcpy((char *) &tmp, self->ob_item + p, (size_t) m);
             cnt += popcnt_64(tmp);
         }
         cnt += count(self, 8 * (p + m), b);
